+++ c/tskit/trees.h
@@ -53,6 +53,7 @@ extern "C" {
 #define TSK_STAT_SPAN_NORMALISE          (1 << 11)
 #define TSK_STAT_ALLOW_TIME_UNCALIBRATED (1 << 12)
 #define TSK_STAT_PAIR_NORMALISE          (1 << 13)
+#define TSK_STAT_NONCENTRED              (1 << 14)
 
 /* Options for map_mutations */
 #define TSK_MM_FIXED_ANCESTRAL_STATE (1 << 0)